productos en home y elimino forms de la plantilla
