--- TP-WebForm-Equipo-20/Productos.aspx
@@ -1,150 +1,0 @@
-﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="Productos.aspx.cs" Inherits="TP_WebForm_Equipo_20.Productos" %>
-
-<asp:Content ID="Content1" ContentPlaceHolderID="MainContent" runat="server">
-    <style>
-        .card {
-            border: 1px solid #ccc;
-            border-radius: 8px;
-            overflow: hidden;
-        }
-
-        .card-img-top {
-            width: 200px;
-            height: 200px;
-        }
-
-        .card-body {
-            padding: 20px;
-        }
-
-        .card-title {
-            font-size: 18px;
-            font-weight: bold;
-            margin-bottom: 15px;
-        }
-
-        .carousel-control-prev-icon, .carousel-control-next-icon {
-            background-color: black;
-        }
-
-        .card-text {
-            margin-bottom: 10px;
-        }
-
-        .btn-primary {
-            background-color: #007bff;
-            color: #fff;
-            border: none;
-            padding: 10px 20px;
-            border-radius: 5px;
-            cursor: pointer;
-        }
-
-            .btn-primary:hover {
-                background-color: #0056b3;
-            }
-
-        a {
-            display: block;
-            text-align: center;
-            margin-top: 10px;
-            color: #007bff;
-            text-decoration: none;
-        }
-    </style>
-
-    <h1>Productos </h1>
-    <div class="row">
-        <div class="col-6">
-            <div class="mb-3">
-                <asp:Label Text="Filtrar" runat="server" />
-                <asp:TextBox ID="txtFiltro" runat="server" CssClass="form-control" AutoPostBack="true" OnTextChanged="txtFiltro_TextChanged" />
-            </div>
-        </div>
-        <div class="col-6" style="display: flex; flex-direction: column; justify-content: flex-end;">
-            <div class="mb-3">
-                <asp:CheckBox ID="chkFiltroAvanzado" Text="Filtro Avanzado" AutoPostBack="true" OnCheckedChanged="chkFiltroAvanzado_CheckedChanged" runat="server" />
-            </div>
-        </div>
-
-        <%if (chkFiltroAvanzado.Checked)
-            {%>
-        <div class="row">
-            <div class="col-3">
-                <div class="mb-3">
-                    <asp:Label ID="lblCampo" Text="Campo" runat="server" />
-                    <asp:DropDownList ID="ddlCampo" runat="server" CssClass="form-control" AutoPostBack="true" OnSelectedIndexChanged="ddlCampo_SelectedIndexChanged">
-                        <asp:ListItem Text="Precio" />
-                        <asp:ListItem Text="Nombre" />
-                        <asp:ListItem Text="Descripcion" />
-                    </asp:DropDownList>
-                </div>
-            </div>
-            <div class="col-3">
-                <div class="mb-3">
-                    <asp:Label ID="lblCriterio" Text="Criterio" runat="server" />
-                    <asp:DropDownList ID="ddlCriterio" runat="server" CssClass="form-control" AutoPostBack="true">
-                        <asp:ListItem Text="Igual a" />
-                        <asp:ListItem Text="Mayor a" />
-                        <asp:ListItem Text="Menor a" />
-                    </asp:DropDownList>
-                </div>
-            </div>
-            <div class="col-3">
-                <div class="mb-3">
-                    <asp:Label ID="lblFiltroAvanzado" Text="Filtro" runat="server" />
-                    <asp:TextBox ID="txtFiltroAvanzado" CssClass="form-control" runat="server" />
-                </div>
-                <asp:Label ID="lblValidarFiltro" Text="" runat="server" ForeColor="Red" />
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-3">
-                <div class="mb-3">
-                    <asp:Button ID="btnBuscar" Text="Buscar" runat="server" CssClass="btn btn-primary" OnClick="btnBuscar_Click" />
-                </div>
-            </div>
-        </div>
-        <% } %>
-    </div>
-    <div class="row row-cols-1 row-cols-md-3 g-4">
-        <asp:Repeater ID="repRepeater" runat="server">
-            <ItemTemplate>
-                <div class="col">
-                    <div class="card">
-                        <div id="carouselExampleControls_<%# Container.ItemIndex %>" class="carousel slide" data-bs-ride="carousel">
-                            <div class="carousel-inner">
-                                <asp:Repeater ID="rptImages" runat="server" DataSource='<%# Eval("Imagenes") %>'>
-                                    <ItemTemplate>
-                                       <%-- <% listaImagenes= ((List<Imagen>)Eval("Imagenes")) %>--%>
-                                        <div class="carousel-item <%# Container.ItemIndex == 0 ? "active" : "" %>">
-                                            <img src='<%# Eval("UrlImagen") %>' class="d-block mx-auto card-img-top" alt="..." onerror="this.onerror=null; this.src='https://img.freepik.com/vector-premium/no-hay-foto-disponible-icono-vector-simbolo-imagen-predeterminado-imagen-proximamente-sitio-web-o-aplicacion-movil_87543-10615.jpg?w=826';">
-                                        </div>
-                                    </ItemTemplate>
-                                </asp:Repeater>
-                            </div>
-                             <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleControls_<%# Container.ItemIndex %>" data-bs-slide="prev">
-                                <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                                <span class="visually-hidden">Previous</span>
-                            </button>
-                            <button class="carousel-control-next" type="button" data-bs-target="#carouselExampleControls_<%# Container.ItemIndex %>" data-bs-slide="next">
-                                <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                                <span class="visually-hidden">Next</span>
-                            </button>
-                        </div>
-                        <div class="card-body">
-                            <h5 class="card-title"><%#Eval("Nombre") %></h5>
-                            <p class="card-text">Código: <%#Eval("Codigo") %></p>
-                            <p class="card-text">Descripción: <%#Eval("Descripcion") %></p>
-                            <p class="card-text">Marca: <%#Eval("Marca.Descripcion") %></p>
-                            <p class="card-text">Categoría: <%#Eval("Categoria.Descripcion") %></p>
-                            <p class="card-text">Precio: $<%#Eval("Precio") %></p>
-                            <a href="DetalleArticulo.aspx?id=<%#Eval("Id") %>" class="btn btn-primary">Ver Detalle</a>
-                            <a href="Carrito.aspx?id=<%#Eval("Id") %>" class="btn btn-primary">Agregar al Carrito</a>
-                        </div>
-                    </div>
-                </div>
-            </ItemTemplate>
-        </asp:Repeater>
-    </div>
-</asp:Content>
