+++ TP-WebForm-Equipo-20/Site.Master
@@ -21,7 +21,7 @@
                 </button>
                 <div class="collapse navbar-collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
-                        <li class="nav-item"><a class="nav-link" runat="server" href="~/Productos">Productos</a></li>
+                        <li class="nav-item"><a class="nav-link" runat="server" href="~/Default">Productos</a></li>
                         <li class="nav-item"><a class="nav-link" runat="server" href="~/Carrito">Carrito</a></li>
                     </ul>
                 </div>
